+++ untitled/src/com/company/Main.java
@@ -1,0 +1,8 @@
+package com.company;
+
+public class Main {
+
+    public static void main(String[] args) {
+
+    }
+}
